Enum 위치 변경, CI 임시 제거
