+++ Moyoy-Api/src/main/java/com/moyoy/api/auth/security/component/CustomAuthenticationEntryPoint.java
@@ -12,7 +12,7 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import com.moyoy.api.support.util.ErrorResponseWriter;
-import com.moyoy.common.exception.auth.AuthErrorCode;
+import com.moyoy.core.support.error.auth.AuthErrorCode;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
