+++ Moyoy-Core/src/main/java/com/moyoy/core/domain/follow/implement/GithubFollowRelationReader.java
@@ -10,7 +10,7 @@ import org.springframework.cache.annotation.Cacheable;
 import org.springframework.cache.annotation.Caching;
 import org.springframework.stereotype.Component;
 
-import com.moyoy.common.exception.user.UserNotFoundException;
+import com.moyoy.core.support.error.user.UserNotFoundException;
 import com.moyoy.core.domain.github.GithubApiLimitChecker;
 import com.moyoy.core.domain.github.GithubOAuthTokenReader;
 import com.moyoy.core.domain.user.implement.UserReader;
