+++ Moyoy-Core/src/main/java/com/moyoy/core/domain/user/implement/Role.java
@@ -1,4 +1,4 @@
-package com.moyoy.common.enums;
+package com.moyoy.core.domain.user.implement;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
