+++ Moyoy-Core/src/main/java/com/moyoy/core/domain/user/implement/UserReader.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Component;
 
-import com.moyoy.common.exception.user.UserNotFoundException;
+import com.moyoy.core.support.error.user.UserNotFoundException;
 import com.moyoy.infra.database.domain.user.UserCountAndLastId;
 import com.moyoy.infra.database.domain.user.UserEntity;
 import com.moyoy.infra.database.domain.user.UserRepository;
