+++ Moyoy-Core/src/main/java/com/moyoy/core/support/error/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package com.moyoy.common.exception;
+package com.moyoy.core.support.error;
 
 public interface BaseErrorCode {
