+++ Moyoy-Core/src/main/java/com/moyoy/core/support/error/github/GithubErrorCode.java
@@ -1,9 +1,9 @@
-package com.moyoy.common.exception.github;
+package com.moyoy.core.support.error.github;
 
 import static com.moyoy.common.constant.MoyoConstants.*;
 
-import com.moyoy.common.exception.BaseErrorCode;
-import com.moyoy.common.exception.ErrorReason;
+import com.moyoy.core.support.error.BaseErrorCode;
+import com.moyoy.core.support.error.ErrorReason;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
