+++ Moyoy-Core/src/main/java/com/moyoy/core/support/error/pr_review/PrReviewErrorCode.java
@@ -1,10 +1,10 @@
-package com.moyoy.common.exception.pr_review;
+package com.moyoy.core.support.error.pr_review;
 
 import static com.moyoy.common.constant.MoyoConstants.FORBIDDEN;
 import static com.moyoy.common.constant.MoyoConstants.NOT_FOUND;
 
-import com.moyoy.common.exception.BaseErrorCode;
-import com.moyoy.common.exception.ErrorReason;
+import com.moyoy.core.support.error.BaseErrorCode;
+import com.moyoy.core.support.error.ErrorReason;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
