+++ Moyoy-Core/src/main/java/com/moyoy/core/support/error/user/UserErrorCode.java
@@ -1,12 +1,12 @@
-package com.moyoy.common.exception.user;
+package com.moyoy.core.support.error.user;
 
 import static com.moyoy.common.constant.MoyoConstants.*;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import com.moyoy.common.exception.BaseErrorCode;
-import com.moyoy.common.exception.ErrorReason;
+import com.moyoy.core.support.error.BaseErrorCode;
+import com.moyoy.core.support.error.ErrorReason;
 
 @Getter
 @RequiredArgsConstructor
