+++ Moyoy-Infra/src/main/java/com/moyoy/infra/database/domain/ranking/RankingRepository.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 
-import com.moyoy.common.enums.RankingPeriod;
 
 public interface RankingRepository {
 
@@ -14,7 +13,7 @@ public interface RankingRepository {
 	void saveAll(List<RankingEntity> rankings);
 	Optional<RankingEntity> findById(Long userId);
 
-	Slice<RankingEntity> findAll(RankingPeriod duration, Pageable pageable);
+	Slice<RankingEntity> findAll(String duration, Pageable pageable);
 
-	Slice<RankingEntity> findFollowingUserRankings(List<Integer> followingUserIds, RankingPeriod rankingPeriod, Pageable pageable);
+	Slice<RankingEntity> findFollowingUserRankings(List<Integer> followingUserIds, String rankingPeriod, Pageable pageable);
 }
